fix(autocomplete): refactor conditions and streamline logic
