+++ .changeset/calm-berries-sleep.md
@@ -1,0 +1,6 @@
+---
+"@yamada-ui/autocomplete": patch
+"@yamada-ui/select": patch
+---
+
+Fixed a bug where `value` was not reset.
