+++ .changeset/purple-spies-collect.md
@@ -1,0 +1,5 @@
+---
+"@yamada-ui/autocomplete": patch
+---
+
+Fixed a bug where the input value was not reset when `allowCreate` was `true`.
