+++ packages/components/select/src/index.ts
@@ -5,4 +5,5 @@ export type { OptionProps } from "./option"
 export { OptionGroup } from "./option-group"
 export type { OptionGroupProps } from "./option-group"
 export { Select } from "./select"
-export type { SelectItem, SelectProps } from "./select"
+export type { SelectProps } from "./select"
+export type { SelectItem } from "./use-select"
