+++ packages/components/select/src/option.tsx
@@ -1,10 +1,11 @@
 import type { CSSUIObject, HTMLUIProps } from "@yamada-ui/core"
 import type { FC, ReactElement } from "react"
-import type { UseSelectOptionProps } from "./use-select"
+import type { UseSelectOptionProps } from "./use-select-option"
 import { forwardRef, ui } from "@yamada-ui/core"
 import { cx } from "@yamada-ui/utils"
 import { useId } from "react"
-import { useSelectContext, useSelectOption } from "./use-select"
+import { useSelectContext } from "./use-select"
+import { useSelectOption } from "./use-select-option"
 
 interface OptionOptions {
   icon?: null | ReactElement
