+++ packages/components/select/src/use-select-option.ts
@@ -11,21 +11,13 @@ import {
   isUndefined,
   mergeRefs,
 } from "@yamada-ui/utils"
-import { useCallback, useId, useRef } from "react"
-import {
-  useAutocompleteContext,
-  useAutocompleteDescendant,
-} from "./autocomplete-context"
+import { useCallback, useRef } from "react"
+import { useSelectContext, useSelectDescendant } from "./use-select"
 
 const isTargetOption = (target: EventTarget | null): boolean =>
   isHTMLElement(target) && !!target.getAttribute("role")?.startsWith("option")
 
-export interface UseAutocompleteOptionProps
-  extends Omit<HTMLUIProps<"li">, "children" | "value"> {
-  /**
-   * The label of the autocomplete option.
-   */
-  children?: string
+export interface UseSelectOptionProps extends Omit<HTMLUIProps<"li">, "value"> {
   /**
    * If `true`, the list element will be closed when selected.
    *
@@ -33,37 +25,38 @@ export interface UseAutocompleteOptionProps
    */
   closeOnSelect?: boolean
   /**
-   * If `true`, the autocomplete option will be disabled.
+   * If `true`, the select option will be disabled.
    *
    * @default false
    */
   isDisabled?: boolean
   /**
-   * If `true`, the autocomplete option will be focusable.
+   * If `true`, the select option will be focusable.
    *
    * @default false
    */
   isFocusable?: boolean
   /**
-   * The value of the autocomplete option.
+   * The value of the select option.
    */
   value?: string
 }
 
-export const useAutocompleteOption = (props: UseAutocompleteOptionProps) => {
+export const useSelectOption = (props: UseSelectOptionProps) => {
   const {
     closeOnSelect,
+    fieldRef,
     focusedIndex,
-    inputRef,
     omitSelectedValues,
+    placeholder,
+    placeholderInOptions,
     setFocusedIndex,
     value,
     optionProps,
     onChange,
     onClose,
     onFocusNext,
-  } = useAutocompleteContext()
-  const id = useId()
+  } = useSelectContext()
 
   let {
     children,
@@ -79,43 +72,53 @@ export const useAutocompleteOption = (props: UseAutocompleteOptionProps) => {
 
   const itemRef = useRef<HTMLLIElement>(null)
 
-  const { descendants, index, register } = useAutocompleteDescendant({
+  const { descendants, index, register } = useSelectDescendant({
     disabled: trulyDisabled,
   })
 
   const values = descendants.values()
   const frontValues = values.slice(0, index)
+  const hasPlaceholder = !!placeholder && placeholderInOptions
+  const isPlaceholder = hasPlaceholder && index === 0
   const isMulti = isArray(value)
 
-  if (isUndefined(optionValue)) {
+  if (!isPlaceholder && isUndefined(optionValue)) {
     if (isString(children) || isNumber(children)) {
       optionValue = children.toString()
     } else {
       console.warn(
         `${
-          !isMulti ? "Autocomplete" : "MultiAutocomplete"
+          !isMulti ? "Select" : "MultiSelect"
         }: Cannot infer the option value of complex children. Pass a \`value\` prop or use a plain string as children to <Option>.`,
       )
     }
   }
 
-  const isDuplicated = frontValues.some(
-    ({ node }) => node.dataset.value === (optionValue ?? ""),
-  )
+  if (hasPlaceholder && index > 0 && !optionValue) {
+    console.warn(
+      `${
+        !isMulti ? "Select" : "MultiSelect"
+      }: If placeholders are present, All options must be set value. If want to set an empty value, either don't set the placeholder or set 'placeholderInOptions' to false.`,
+    )
+  }
+
+  const isDuplicated = !isMulti
+    ? frontValues.some(({ node }) => node.dataset.value === (optionValue ?? ""))
+    : false
   const isSelected =
     !isDuplicated &&
     (!isMulti
       ? (optionValue ?? "") === value
       : value.includes(optionValue ?? ""))
-  const isTarget = "target" in (itemRef.current?.dataset ?? {})
   const isFocused = index === focusedIndex
 
   const onClick = useCallback(
     (ev: MouseEvent<HTMLLIElement>) => {
+      ev.preventDefault()
       ev.stopPropagation()
 
       if (isDisabled || !isTargetOption(ev.currentTarget)) {
-        if (inputRef.current) inputRef.current.focus()
+        if (fieldRef.current) fieldRef.current.focus()
 
         return
       }
@@ -124,25 +127,25 @@ export const useAutocompleteOption = (props: UseAutocompleteOptionProps) => {
 
       onChange(optionValue ?? "")
 
-      if (inputRef.current) inputRef.current.focus()
+      if (fieldRef.current) fieldRef.current.focus()
 
       if (customCloseOnSelect ?? closeOnSelect) onClose()
 
-      if (omitSelectedValues) onFocusNext(index)
+      if (omitSelectedValues) onFocusNext()
     },
     [
-      isDuplicated,
-      onFocusNext,
-      omitSelectedValues,
       isDisabled,
-      optionValue,
+      isDuplicated,
       setFocusedIndex,
       index,
       onChange,
+      optionValue,
+      fieldRef,
       customCloseOnSelect,
       closeOnSelect,
       onClose,
-      inputRef,
+      omitSelectedValues,
+      onFocusNext,
     ],
   )
 
@@ -161,33 +164,28 @@ export const useAutocompleteOption = (props: UseAutocompleteOptionProps) => {
       }
 
       return {
-        id,
         ref: mergeRefs(itemRef, ref, register),
         role: "option",
         ...computedProps,
         ...props,
-        style:
-          !isTarget || (omitSelectedValues && isSelected) ? style : undefined,
-        "aria-checked": isSelected,
+        style: omitSelectedValues && isSelected ? style : undefined,
+        "aria-checked": ariaAttr(isSelected),
         "aria-disabled": ariaAttr(isDisabled),
         "data-disabled": dataAttr(isDisabled),
         "data-duplicated": dataAttr(isDuplicated),
         "data-focus": dataAttr(isFocused),
-        "data-target": dataAttr(true),
         "data-value": optionValue ?? "",
         tabIndex: -1,
         onClick: handlerAll(computedProps.onClick, props.onClick, onClick),
       }
     },
     [
-      id,
-      isDuplicated,
       optionValue,
       computedProps,
       isDisabled,
       isFocused,
+      isDuplicated,
       isSelected,
-      isTarget,
       omitSelectedValues,
       onClick,
       register,
@@ -203,74 +201,4 @@ export const useAutocompleteOption = (props: UseAutocompleteOptionProps) => {
   }
 }
 
-export type UseAutocompleteOptionReturn = ReturnType<
-  typeof useAutocompleteOption
->
-
-export const useAutocompleteCreate = () => {
-  const { isHit, onCreate } = useAutocompleteContext()
-
-  const getCreateProps: PropGetter<"li"> = useCallback(
-    (props = {}, ref = null) => {
-      const style: CSSProperties = {
-        border: "0px",
-        clip: "rect(0px, 0px, 0px, 0px)",
-        height: "1px",
-        margin: "-1px",
-        overflow: "hidden",
-        padding: "0px",
-        position: "absolute",
-        whiteSpace: "nowrap",
-        width: "1px",
-      }
-
-      return {
-        ref,
-        ...props,
-        style: isHit ? style : undefined,
-        "data-focus": dataAttr(!isHit),
-        tabIndex: -1,
-        onClick: handlerAll(props.onClick, onCreate),
-      }
-    },
-    [isHit, onCreate],
-  )
-
-  return { getCreateProps }
-}
-
-export type UseAutocompleteCreateReturn = ReturnType<
-  typeof useAutocompleteCreate
->
-
-export const useAutocompleteEmpty = () => {
-  const { isEmpty, isHit } = useAutocompleteContext()
-
-  const getEmptyProps: PropGetter<"li"> = useCallback(
-    (props = {}, ref = null) => {
-      const style: CSSProperties = {
-        border: "0px",
-        clip: "rect(0px, 0px, 0px, 0px)",
-        height: "1px",
-        margin: "-1px",
-        overflow: "hidden",
-        padding: "0px",
-        position: "absolute",
-        whiteSpace: "nowrap",
-        width: "1px",
-      }
-
-      return {
-        ref,
-        ...props,
-        style: isHit && !isEmpty ? style : undefined,
-        tabIndex: -1,
-      }
-    },
-    [isHit, isEmpty],
-  )
-
-  return { getEmptyProps }
-}
-
-export type UseAutocompleteEmptyReturn = ReturnType<typeof useAutocompleteEmpty>
+export type UseSelectOptionReturn = ReturnType<typeof useSelectOption>
